5.11 - fix: delete corresponding blog from a user's blog list when deleting a blog
- delete blogs id from the users list of blogs
- allow old format blogs with no associated user to be deleted
- send 404 response when a blog that does not exist on server is requested to be deleted

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const userExtractor = require("../utils/userExtractor")
+const User = require("../models/user");
+const userExtractor = require("../utils/userExtractor");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { blogs: 0 });
@@ -24,8 +25,23 @@ blogsRouter.delete("/:id", userExtractor, async (request, response) => {
   const user = request.user;
   const blog = await Blog.findById(request.params.id);
 
+  if (!blog) {
+    let error = new Error(`the blog with id - ${request.params.id} was not found on the server`);
+    error.name = "MissingData";
+    throw error;
+  }
+
+  //Legacy Support
+  //allow blogs which dont have an assigned user to be deleted
+  if (!blog.user) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(200).end();
+    return;
+  }
+
   if (user.id.toString() === blog.user.toString()) {
     await Blog.findByIdAndDelete(request.params.id);
+    await User.findByIdAndUpdate(user.id, { $pull: { blogs: blog.id } }, { new: true });
     response.status(200).end();
   } else {
     let error = new Error("a blog can only be deleted by its author");
